requirements.txt
# Core FastAPI and dependencies (from LLM Platform)
fastapi>=0.104.0
httpx[brotli]
requests>=2.31.0
openai>=1.3.0
passlib>=1.7.4
psycopg2-binary>=2.9.9
pydantic>=2.4.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
pytest>=7.4.3
pytest-asyncio>=0.23.0
pytest-cov>=6.0.0
pytest-xdist>=3.5.0
slowapi>=0.1.4
sqlalchemy>=2.0.23
uvicorn>=0.24.0
aiohttp>=3.9.1
websockets>=12.0  # WebSocket client for Gateway proxy to KB Service
alembic>=1.12.1
cryptography>=41.0.0
anthropic>=0.57.0
bcrypt>=4.0.1
email-validator>=2.0.0
mcp[cli]>=0.1.0
PyJWT>=2.8.0
supabase
redis>=4.0.0
pgvector>=0.2.0
pillow>=10.0.0
aiofiles>=23.0.0
numpy>=1.24.0
aioredis>=2.0.0

# Additional Gaia Platform dependencies
nats-py>=2.3.0
mcp-agent>=0.1.0
fastmcp>=0.2.0  # FastMCP HTTP server for Claude Code integration

# KB integration dependencies
python-frontmatter>=1.1.0
asyncpg>=0.29.0

# Semantic search dependencies (pgvector-based)
sentence-transformers>=2.2.2  # For generating embeddings (all-MiniLM-L6-v2)
unstructured[all-docs]>=0.10.0
